% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2intmat.R
\name{df2intmat}
\alias{df2intmat}
\title{Converts data frame to integer matrix}
\usage{
df2intmat(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
The conversion is done by converting to characters, then a factor and then integer
}
\examples{
x <- mtcars[1:4, 1:5]
x
df2intmat(x)

}
