% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{entropy}
\alias{entropy}
\title{Entropy using `log()` (natural logarithm)}
\usage{
entropy(p)
}
\arguments{
\item{p}{Vector of probabilities that constitutes a distribution (all >= 0 and <= 1 and sums to 1)}
}
\description{
Entropy using `log()` (natural logarithm)
}
