% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{entropy10}
\alias{entropy10}
\title{Entropy using `log10()`}
\usage{
entropy10(p)
}
\arguments{
\item{p}{Vector of probabilities that constitutes a distribution (all >= 0 and <= 1 and sums to 1)}
}
\description{
Entropy using `log10()`
}
