% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{entropy2}
\alias{entropy2}
\title{Entropy using `log2()`}
\usage{
entropy2(p)
}
\arguments{
\item{p}{Vector of probabilities that constitutes a distribution (all >= 0 and <= 1 and sums to 1)}
}
\description{
Entropy using `log2()`
}
