% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{entropy_cond10}
\alias{entropy_cond10}
\title{Conditional entropy of $X$ given $Y$ using `log10()`}
\usage{
entropy_cond10(d, idx_x, idx_y)
}
\arguments{
\item{d}{Integer matrix of data where each row is a single observation}

\item{idx_x}{Indicies (integer vector) for variable set $X$ (in 1 to `ncol(d)`)}

\item{idx_y}{Indicies (integer vector) for variable set $Y$ (in 1 to `ncol(d)`)}
}
\description{
Conditional entropy of $X$ given $Y$ using `log10()`
}
