% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{entropyE}
\alias{entropyE}
\title{Entropy using `log()` (natural logarithm)}
\usage{
entropyE(d)
}
\arguments{
\item{d}{Vector or matrix of observations}
}
\description{
Se Introduction vignette for definition.
}
