% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{frequencies_2d}
\alias{frequencies_2d}
\title{Frequencies}
\usage{
frequencies_2d(x, idx_x, idx_y)
}
\arguments{
\item{x}{Integer matrix of data where each row is a single observation}

\item{idx_x}{Indicies (integer vector) for variable set $X$}

\item{idx_y}{Indicies (integer vector) for variable set $Y$}
}
\description{
Similar to `table()` for two sets of variables, $X$ and $Y$, 
but returns a matrix with counts instead
}
