% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{infdist2}
\alias{infdist2}
\title{Shared information distance of $X$ and $Y$ using `log2()`}
\usage{
infdist2(d, idx_x, idx_y)
}
\arguments{
\item{d}{Integer matrix of data where each row is a single observation}

\item{idx_x}{Indicies (integer vector) for variable set $X$ (in 1 to `ncol(d)`)}

\item{idx_y}{Indicies (integer vector) for variable set $Y$ (in 1 to `ncol(d)`)}
}
\description{
Shared information distance of $X$ and $Y$ using `log2()`
}
