% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mutual_information}
\alias{mutual_information}
\title{Mutual information using `log()` (natural logarithm)}
\usage{
mutual_information(ps)
}
\arguments{
\item{Matrix}{of probabilities obtained from [frequencies_2d()] and [normalise_2d()]}
}
\description{
Mutual information using `log()` (natural logarithm)
}
