% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{mutual_information_implicit}
\alias{mutual_information_implicit}
\title{Mutual information using `log()` (natural logarithm)}
\usage{
mutual_information_implicit(x, idx_x, idx_y)
}
\arguments{
\item{x}{Integer matrix of data where each row is a single observation}

\item{idx_x}{Indicies (integer vector) for variable set $X$}

\item{idx_y}{Indicies (integer vector) for variable set $Y$}
}
\description{
Mutual information using `log()` (natural logarithm)
}
